Updated logic to use api route instead of client-side fetching for public profile and renamed Profile page as Account
