+++ components/account/friends.tsx
